requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
flake8
python-dotenv>=0.5.1
timm
torch
torchvision
wandb
pandas
Pillow
pytest
hydra-core
isort
black
fastapi
uvicorn[standard]
python-multipart